migrate *broadcast_ methods to namedarray
