+++ xarray/namedarray/utils.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import importlib
 import sys
 import warnings
-from collections.abc import Hashable, Iterable, Iterator, Mapping
+from collections.abc import Hashable, Iterable, Iterator, Mapping, MutableSet
 from functools import lru_cache
 from typing import TYPE_CHECKING, Any, TypeVar, cast
 
@@ -222,3 +222,45 @@ class ReprObject:
         from dask.base import normalize_token
 
         return normalize_token((type(self), self._value))
+
+
+class OrderedSet(MutableSet[T]):
+    """A simple ordered set.
+
+    The API matches the builtin set, but it preserves insertion order of elements, like
+    a dict. Note that, unlike in an OrderedDict, equality tests are not order-sensitive.
+    """
+
+    _d: dict[T, None]
+
+    __slots__ = ("_d",)
+
+    def __init__(self, values: Iterable[T] | None = None):
+        self._d = {}
+        if values is not None:
+            self.update(values)
+
+    # Required methods for MutableSet
+
+    def __contains__(self, value: Hashable) -> bool:
+        return value in self._d
+
+    def __iter__(self) -> Iterator[T]:
+        return iter(self._d)
+
+    def __len__(self) -> int:
+        return len(self._d)
+
+    def add(self, value: T) -> None:
+        self._d[value] = None
+
+    def discard(self, value: T) -> None:
+        del self._d[value]
+
+    # Additional methods
+
+    def update(self, values: Iterable[T]) -> None:
+        self._d.update(dict.fromkeys(values))
+
+    def __repr__(self) -> str:
+        return f"{type(self).__name__}({list(self)!r})"
